src/main/java/com/example/study/real_world/BankStatementCSVParser.java
package com.example.study.real_world;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.util.ArrayList;
import java.util.List;

public class BankStatementCSVParser implements BankStatementParser {

    public static final DateTimeFormatter DATE_PATTERN = DateTimeFormatter.ofPattern("dd-MM-yyyy");

    @Override
    public BankTransaction parseFrom(String line) {
        final String[] columns = line.split(",");

        final LocalDate date = LocalDate.parse(columns[0], DATE_PATTERN);
        final double amount = Double.parseDouble(columns[1]);
        final String description = columns[2];

        return new BankTransaction(date, amount, description);
    }

    @Override
    public List<BankTransaction> parserLinesFrom(List<String> lines) {
        final List<BankTransaction> bankTransactionList = new ArrayList<>();
        for (final String line : lines) {
            bankTransactionList.add(parseFrom(line));
        }
        return bankTransactionList;
    }
}
